Fixed crash when browsing for audio or image files
Alice would crash if you navigated through the Alice app folder on a
mac while browsing for sounds or images.

Fixed a check in the file browser code that didn’t account for null
extensions on files (which happens when a file doesn’t have an
extension)

Add this sort of check to a few other places as well.


Former-commit-id: 9103e7b44a34353300e123f035bd65cd661149f1