change how we get icons by size
This is the big change- this moves 99% of the logic about the final size of the icon into the factories, instead of the innocent code that just wants an icon.

I left the previous behaviour of getIconExactSize in case we find a case where we really do want to scale something to an exact size.

I suspect that most code will just want to call getIconToFit, which I could have left as getIcon, but I was feeling wordy, I guess. Also, now git blame will correctly call me out if something goes sideways...