simplify TransformationUtilities
There are now three functions offered by transformationUtilities

transformToAbsolute
transformFromAbsolute
transformTo

and we offer them for Vector4, Vector3, and Point3.
(not all functions are used for all types, but it is not unreasonable to think they might be)

all of them return a brand new object, created bespoke for the calling function.